more refactor, tests pass, creating pvm classes
