+++ src/apis/avm/index.ts
@@ -1,10 +1,15 @@
 export * from './api';
+export * from './basetx';
 export * from './constants';
+export * from './createassettx';
 export * from './credentials';
+export * from './exporttx';
+export * from './importtx';
 export * from './initialstates';
 export * from './inputs';
 export * from './keychain';
 export * from './minterset';
+export * from './operationtx';
 export * from './ops';
 export * from './outputs';
 export * from './tx';
