+++ src/apis/platformvm/index.ts
@@ -1,6 +1,9 @@
 export * from './api';
+export * from './basetx';
 export * from './constants';
 export * from './credentials';
+export * from './exporttx';
+export * from './importtx';
 export * from './inputs';
 export * from './keychain';
 export * from './outputs';
